examples/logs-slack-alerts/function_source/requirements.txt
google-cloud-bigquery==1.14.0
requests==2.21.0